pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
dependencies = [
    "pytest",
    "black",
    "isort",
    "flake8",
    "pre-commit"
]
build-backend = "setuptools.build_meta"

[project]
name = "ic-ocr-catalogger"
version = "0.2.0"
description = "An python OpenCV UI that performs live OCR chips, with an integrated catalog."
authors = [
    { name = "Rose Hooper", email = "rose@rosehooper.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
license = { text = "MIT" }
dependencies = [
    "opencv-python~=4.11.0.86",
    "opencv-contrib-python~=4.11.0.86",
    "numpy~=2.2.4",
    "pillow~=11.1.0",
    "more-itertools~=10.6.0",
    "ocrmac~=1.0.0"
]

[tool.setuptools]
packages = ["ic_ocr_cataloger"]

[tool.setuptools.package-data]
"ic_ocr_cataloger" = ["data/*.txt", "data/*.tsv", "data/*.json", "default_config.ini"]

[project.scripts]
"icocr" = "ic_ocr_cataloger.__main__:main"

[tool.isort]
profile = "black"
